Clean up, restructure events.ts
